u64 metrics

+++ bee-protocol/src/protocol/metrics.rs
@@ -1,27 +1,27 @@
 use std::sync::atomic::{
-    AtomicU32,
+    AtomicU64,
     Ordering,
 };
 
 #[derive(Default)]
 pub struct ProtocolMetrics {
-    invalid_transactions_received: AtomicU32,
-    stale_transactions_received: AtomicU32,
-    random_transactions_received: AtomicU32,
-    new_transactions_received: AtomicU32,
-    known_transactions_received: AtomicU32,
+    invalid_transactions_received: AtomicU64,
+    stale_transactions_received: AtomicU64,
+    random_transactions_received: AtomicU64,
+    new_transactions_received: AtomicU64,
+    known_transactions_received: AtomicU64,
 
-    invalid_messages_received: AtomicU32,
+    invalid_messages_received: AtomicU64,
 
-    milestone_request_received: AtomicU32,
-    transaction_broadcast_received: AtomicU32,
-    transaction_request_received: AtomicU32,
-    heartbeat_received: AtomicU32,
+    milestone_request_received: AtomicU64,
+    transaction_broadcast_received: AtomicU64,
+    transaction_request_received: AtomicU64,
+    heartbeat_received: AtomicU64,
 
-    milestone_request_sent: AtomicU32,
-    transaction_broadcast_sent: AtomicU32,
-    transaction_request_sent: AtomicU32,
-    heartbeat_sent: AtomicU32,
+    milestone_request_sent: AtomicU64,
+    transaction_broadcast_sent: AtomicU64,
+    transaction_request_sent: AtomicU64,
+    heartbeat_sent: AtomicU64,
 }
 
 impl ProtocolMetrics {
@@ -31,115 +31,115 @@ impl ProtocolMetrics {
 }
 
 impl ProtocolMetrics {
-    pub fn invalid_transactions_received(&self) -> u32 {
+    pub fn invalid_transactions_received(&self) -> u64 {
         self.invalid_transactions_received.load(Ordering::Relaxed)
     }
 
-    pub fn invalid_transactions_received_inc(&self) -> u32 {
+    pub fn invalid_transactions_received_inc(&self) -> u64 {
         self.invalid_transactions_received.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn stale_transactions_received(&self) -> u32 {
+    pub fn stale_transactions_received(&self) -> u64 {
         self.stale_transactions_received.load(Ordering::Relaxed)
     }
 
-    pub fn stale_transactions_received_inc(&self) -> u32 {
+    pub fn stale_transactions_received_inc(&self) -> u64 {
         self.stale_transactions_received.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn random_transactions_received(&self) -> u32 {
+    pub fn random_transactions_received(&self) -> u64 {
         self.random_transactions_received.load(Ordering::Relaxed)
     }
 
-    pub fn random_transactions_received_inc(&self) -> u32 {
+    pub fn random_transactions_received_inc(&self) -> u64 {
         self.random_transactions_received.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn new_transactions_received(&self) -> u32 {
+    pub fn new_transactions_received(&self) -> u64 {
         self.new_transactions_received.load(Ordering::Relaxed)
     }
 
-    pub fn new_transactions_received_inc(&self) -> u32 {
+    pub fn new_transactions_received_inc(&self) -> u64 {
         self.new_transactions_received.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn known_transactions_received(&self) -> u32 {
+    pub fn known_transactions_received(&self) -> u64 {
         self.known_transactions_received.load(Ordering::Relaxed)
     }
 
-    pub fn known_transactions_received_inc(&self) -> u32 {
+    pub fn known_transactions_received_inc(&self) -> u64 {
         self.known_transactions_received.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn invalid_messages_received(&self) -> u32 {
+    pub fn invalid_messages_received(&self) -> u64 {
         self.invalid_messages_received.load(Ordering::Relaxed)
     }
 
-    pub fn invalid_messages_received_inc(&self) -> u32 {
+    pub fn invalid_messages_received_inc(&self) -> u64 {
         self.invalid_messages_received.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn milestone_request_received(&self) -> u32 {
+    pub fn milestone_request_received(&self) -> u64 {
         self.milestone_request_received.load(Ordering::Relaxed)
     }
 
-    pub fn milestone_request_received_inc(&self) -> u32 {
+    pub fn milestone_request_received_inc(&self) -> u64 {
         self.milestone_request_received.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn transaction_broadcast_received(&self) -> u32 {
+    pub fn transaction_broadcast_received(&self) -> u64 {
         self.transaction_broadcast_received.load(Ordering::Relaxed)
     }
 
-    pub fn transaction_broadcast_received_inc(&self) -> u32 {
+    pub fn transaction_broadcast_received_inc(&self) -> u64 {
         self.transaction_broadcast_received.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn transaction_request_received(&self) -> u32 {
+    pub fn transaction_request_received(&self) -> u64 {
         self.transaction_request_received.load(Ordering::Relaxed)
     }
 
-    pub fn transaction_request_received_inc(&self) -> u32 {
+    pub fn transaction_request_received_inc(&self) -> u64 {
         self.transaction_request_received.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn heartbeat_received(&self) -> u32 {
+    pub fn heartbeat_received(&self) -> u64 {
         self.heartbeat_received.load(Ordering::Relaxed)
     }
 
-    pub fn heartbeat_received_inc(&self) -> u32 {
+    pub fn heartbeat_received_inc(&self) -> u64 {
         self.heartbeat_received.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn milestone_request_sent(&self) -> u32 {
+    pub fn milestone_request_sent(&self) -> u64 {
         self.milestone_request_sent.load(Ordering::Relaxed)
     }
 
-    pub fn milestone_request_sent_inc(&self) -> u32 {
+    pub fn milestone_request_sent_inc(&self) -> u64 {
         self.milestone_request_sent.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn transaction_broadcast_sent(&self) -> u32 {
+    pub fn transaction_broadcast_sent(&self) -> u64 {
         self.transaction_broadcast_sent.load(Ordering::Relaxed)
     }
 
-    pub fn transaction_broadcast_sent_inc(&self) -> u32 {
+    pub fn transaction_broadcast_sent_inc(&self) -> u64 {
         self.transaction_broadcast_sent.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn transaction_request_sent(&self) -> u32 {
+    pub fn transaction_request_sent(&self) -> u64 {
         self.transaction_request_sent.load(Ordering::Relaxed)
     }
 
-    pub fn transaction_request_sent_inc(&self) -> u32 {
+    pub fn transaction_request_sent_inc(&self) -> u64 {
         self.transaction_request_sent.fetch_add(1, Ordering::SeqCst)
     }
 
-    pub fn heartbeat_sent(&self) -> u32 {
+    pub fn heartbeat_sent(&self) -> u64 {
         self.heartbeat_sent.load(Ordering::Relaxed)
     }
 
-    pub fn heartbeat_sent_inc(&self) -> u32 {
+    pub fn heartbeat_sent_inc(&self) -> u64 {
         self.heartbeat_sent.fetch_add(1, Ordering::SeqCst)
     }
 }
